Use libps and enhance [pid]/stat
* Makefile: Add libps to the $(LIBS).
* proclist.c, proclist.h: Embed the proc server port in a
  ps_context structure. (proclist_make_node): Change to prototype to
  allow for the possibility of error.  Rename to proclist_create_node to
  reflect the change and non-triviality.
* process.c, process.h: Revamp.  Use a full-blown procstat
  structure instead of just the procinfo fetched from the process
  server.  Use the additional data to complement [pid]/stat.
  (process_lookup_pid): Get a ps_context structure instead of a port to
  the process server.
* main.c (root_make_node): Convert to the new interface for
  proclist_create_node.

+++ procfs.c
@@ -87,7 +87,7 @@ procfs_make_ino (struct node *np, const char *filename)
       jrand48 (x);
     }
 
-  return jrand48 (x);
+  return (unsigned long) jrand48 (x);
 }
 
 error_t procfs_get_contents (struct node *np, void **data, size_t *data_len)
